Crate ibc_app_transfer_types

Implementation of the IBC fungible token transfer (ICS-20) data structures.

Modules

error

Defines the token transfer error type

events

Defines all token transfer event types

msgs

Defines the token transfer message type

packet

Contains the PacketData type that defines the structure of token transfers’ packet bytes

proto

Re-exports ICS-20 token transfer proto types from the ibc-proto crate.

Structs

Amount

A type for representing token transfer amounts.

BaseDenom

The “base” of a denomination.

Coin

Coin defines a token with a denomination and an amount.

Memo

Represents the token transfer memo

PrefixedDenom

A type that contains the base denomination for ICS20 and the source tracing information path.

TracePath

A full trace path modelled as a collection of TracePrefixs.

TracePrefix

One hop in a token’s trace, which consists of the port and channel IDs of the sender

U256

Re-exports U256 from primitive-types crate for convenience. Little-endian large integer type 256-bit unsigned integer.

Constants

ACK_SUCCESS_B64

The successful string, used for creating an acknowledgement status. It is equivalent to base64::encode(0x01).

MODULE_ID_STR

Module identifier for the ICS20 application.

PORT_ID_STR

The port identifier that the ICS20 applications typically bind with.

VERSION

ICS20 application current version.

Functions

ack_success_b64

Returns a successful acknowledgement status for the token transfer application.

is_receiver_chain_source

Returns true if the denomination originally came from the receiving chain, and false otherwise.

is_sender_chain_source

Returns true if the denomination originally came from the sender chain, and false otherwise.

Type Aliases

BaseCoin

A Coin type with an unprefixed denomination.

PrefixedCoin

A Coin type with fully qualified PrefixedDenom.

RawCoin